requirements.txt
fastapi
uvicorn
psycopg2
pydantic
python-dotenv
jinaj2